version 0.98 añadido control md5

+++ Jiayu/src/es/jiayu/jiayuid/RomScreen.java
@@ -288,56 +288,65 @@ public class RomScreen extends Activity implements AdapterView.OnItemSelectedLis
                 if (chkCWM.isChecked()) {
 
                     if (controlRoot()) {
-                        Runtime rt = Runtime.getRuntime();
-                        java.lang.Process p = rt.exec("su");
-                        BufferedOutputStream bos = new BufferedOutputStream(
-                                p.getOutputStream());
-                        bos.write(("rm /cache/recovery/extendedcommand\n")
-                                .getBytes());
-                        String fileCWM = "";
-                        if("G4A".equals(modelo) || "S1".equals(modelo)){
-                            fileCWM = this.romseleccionada.replaceAll(Environment.getExternalStorageDirectory().getAbsolutePath(), "/emmc");
-                            bos.write(("echo 'run_program(\"/sbin/umount\",\"/emmc\");' >> /cache/recovery/extendedcommand\n").getBytes());
-                            bos.write(("echo 'run_program(\"/sbin/mount\",\"/emmc\");' >> /cache/recovery/extendedcommand\n").getBytes());
+                        if(Utilidades.checkFileMD5(new File(this.romseleccionada))){
+                            Runtime rt = Runtime.getRuntime();
+                            java.lang.Process p = rt.exec("su");
+                            BufferedOutputStream bos = new BufferedOutputStream(
+                                    p.getOutputStream());
+                            bos.write(("rm /cache/recovery/extendedcommand\n")
+                                    .getBytes());
+                            String fileCWM = "";
+                            if("G4A".equals(modelo) || "S1".equals(modelo)){
+                                fileCWM = this.romseleccionada.replaceAll(Environment.getExternalStorageDirectory().getAbsolutePath(), "/emmc");
+                                bos.write(("echo 'run_program(\"/sbin/umount\",\"/emmc\");' >> /cache/recovery/extendedcommand\n").getBytes());
+                                bos.write(("echo 'run_program(\"/sbin/mount\",\"/emmc\");' >> /cache/recovery/extendedcommand\n").getBytes());
+                            }else{
+                                fileCWM =this.romseleccionada.replaceAll(Environment.getExternalStorageDirectory().getAbsolutePath(), "/sdcard");
+                                bos.write(("echo 'run_program(\"/sbin/umount\",\"/sdcard\");' >> /cache/recovery/extendedcommand\n").getBytes());
+                                bos.write(("echo 'run_program(\"/sbin/mount\",\"/sdcard\");' >> /cache/recovery/extendedcommand\n").getBytes());
+                            }
+                            bos.write(("echo 'install_zip(\"" + fileCWM + "\");\n' >> /cache/recovery/extendedcommand\n").getBytes());
+                                    /*String fileCWM2=this.romseleccionada.replaceAll(Environment.getExternalStorageDirectory().getAbsolutePath(),"/sdcard2");
+                                    bos.write(("echo 'install_zip(\""+ fileCWM2 +"\");' >> /cache/recovery/extendedcommand\n").getBytes());*/
+                            bos.flush();
+                            bos.close();
+                            rebootRecoveryQuestionFlashear();
                         }else{
-                            fileCWM =this.romseleccionada.replaceAll(Environment.getExternalStorageDirectory().getAbsolutePath(), "/sdcard");
-                            bos.write(("echo 'run_program(\"/sbin/umount\",\"/sdcard\");' >> /cache/recovery/extendedcommand\n").getBytes());
-                            bos.write(("echo 'run_program(\"/sbin/mount\",\"/sdcard\");' >> /cache/recovery/extendedcommand\n").getBytes());
+                            Toast.makeText(getBaseContext(),getResources().getString(R.string.msgErrorMD5),Toast.LENGTH_SHORT).show();
                         }
-                        bos.write(("echo 'install_zip(\"" + fileCWM + "\");\n' >> /cache/recovery/extendedcommand\n").getBytes());
-                                /*String fileCWM2=this.romseleccionada.replaceAll(Environment.getExternalStorageDirectory().getAbsolutePath(),"/sdcard2");
-                                bos.write(("echo 'install_zip(\""+ fileCWM2 +"\");' >> /cache/recovery/extendedcommand\n").getBytes());*/
-                        bos.flush();
-                        bos.close();
-                        rebootRecoveryQuestionFlashear();
                     }
                 } else {
                     String application_name = "";
                     try {
-                        application_name = "com.mediatek.updatesystem.UpdateSystem";
-                        Intent intent = new Intent("android.intent.action.MAIN");
-                        List<ResolveInfo> resolveinfo_list = getPackageManager().queryIntentActivities(intent, 0);
-                        boolean existe = false;
-                        for (ResolveInfo info : resolveinfo_list) {
-                            if (info.activityInfo.packageName.equalsIgnoreCase("com.mediatek.updatesystem")) {
-                                if (info.activityInfo.name.equalsIgnoreCase(application_name)) {
-                                    File f = new File(this.romseleccionada);
-                                    if (new File(Environment.getExternalStorageDirectory() + "/update.zip").exists()) {
-                                        new File(Environment.getExternalStorageDirectory() + "/update.zip").delete();
+                        if(Utilidades.checkFileMD5(new File(this.romseleccionada))){
+                            application_name = "com.mediatek.updatesystem.UpdateSystem";
+                            Intent intent = new Intent("android.intent.action.MAIN");
+                            List<ResolveInfo> resolveinfo_list = getPackageManager().queryIntentActivities(intent, 0);
+                            boolean existe = false;
+                            for (ResolveInfo info : resolveinfo_list) {
+                                if (info.activityInfo.packageName.equalsIgnoreCase("com.mediatek.updatesystem")) {
+                                    if (info.activityInfo.name.equalsIgnoreCase(application_name)) {
+                                        File f = new File(this.romseleccionada);
+                                        if (new File(Environment.getExternalStorageDirectory() + "/update.zip").exists()) {
+                                            new File(Environment.getExternalStorageDirectory() + "/update.zip").delete();
 
+                                        }
+
+                                            f.renameTo(new File(Environment.getExternalStorageDirectory() + "/update.zip"));
+                                            Intent launch_intent = new Intent("android.intent.action.MAIN");
+                                            launch_intent.setComponent(new ComponentName(info.activityInfo.packageName, info.activityInfo.name));
+                                            launch_intent.setFlags(Intent.FLAG_ACTIVITY_NEW_TASK);
+                                            this.startActivity(launch_intent);
+                                            existe = true;
+                                            break;
                                     }
-                                    f.renameTo(new File(Environment.getExternalStorageDirectory() + "/update.zip"));
-                                    Intent launch_intent = new Intent("android.intent.action.MAIN");
-                                    launch_intent.setComponent(new ComponentName(info.activityInfo.packageName, info.activityInfo.name));
-                                    launch_intent.setFlags(Intent.FLAG_ACTIVITY_NEW_TASK);
-                                    this.startActivity(launch_intent);
-                                    existe = true;
-                                    break;
                                 }
                             }
-                        }
-                        if (!existe) {
-                            Toast.makeText(getBaseContext(), getResources().getString(R.string.msgIngenieroNoExiste), Toast.LENGTH_SHORT).show();
+                            if (!existe) {
+                                Toast.makeText(getBaseContext(), getResources().getString(R.string.msgIngenieroNoExiste), Toast.LENGTH_SHORT).show();
+                            }
+                        }else{
+                            Toast.makeText(getBaseContext(),getResources().getString(R.string.msgErrorMD5),Toast.LENGTH_SHORT).show();
                         }
                     } catch (ActivityNotFoundException e) {
                         Toast.makeText(getBaseContext(), getResources().getString(R.string.msgGenericError) + application_name+" 155", Toast.LENGTH_SHORT).show();
@@ -353,29 +362,33 @@ public class RomScreen extends Activity implements AdapterView.OnItemSelectedLis
     public void flashZip(){
         try {
             if(aceptadoNoModelo){
-                Runtime rt = Runtime.getRuntime();
-                java.lang.Process p = rt.exec("su");
-                BufferedOutputStream bos = new BufferedOutputStream(
-                        p.getOutputStream());
-                bos.write(("rm /cache/recovery/extendedcommand\n")
-                        .getBytes());
-                String fileCWM="";
-                if("G4A".equals(modelo) || "S1".equals(modelo)){
-                    fileCWM = this.zipseleccionada.replaceAll(Environment.getExternalStorageDirectory().getAbsolutePath(), "/emmc");
-                    bos.write(("echo 'run_program(\"/sbin/umount\",\"/emmc\");' >> /cache/recovery/extendedcommand\n").getBytes());
-                    bos.write(("echo 'run_program(\"/sbin/mount\",\"/emmc\");' >> /cache/recovery/extendedcommand\n").getBytes());
-                }else{
-                    fileCWM =this.zipseleccionada.replaceAll(Environment.getExternalStorageDirectory().getAbsolutePath(), "/sdcard");
-                    bos.write(("echo 'run_program(\"/sbin/umount\",\"/sdcard\");' >> /cache/recovery/extendedcommand\n").getBytes());
-                    bos.write(("echo 'run_program(\"/sbin/mount\",\"/sdcard\");' >> /cache/recovery/extendedcommand\n").getBytes());
-                }
+                if(Utilidades.checkFileMD5(new File(this.zipseleccionada))){
+                    Runtime rt = Runtime.getRuntime();
+                    java.lang.Process p = rt.exec("su");
+                    BufferedOutputStream bos = new BufferedOutputStream(
+                            p.getOutputStream());
+                    bos.write(("rm /cache/recovery/extendedcommand\n")
+                            .getBytes());
+                    String fileCWM="";
+                    if("G4A".equals(modelo) || "S1".equals(modelo)){
+                        fileCWM = this.zipseleccionada.replaceAll(Environment.getExternalStorageDirectory().getAbsolutePath(), "/emmc");
+                        bos.write(("echo 'run_program(\"/sbin/umount\",\"/emmc\");' >> /cache/recovery/extendedcommand\n").getBytes());
+                        bos.write(("echo 'run_program(\"/sbin/mount\",\"/emmc\");' >> /cache/recovery/extendedcommand\n").getBytes());
+                    }else{
+                        fileCWM =this.zipseleccionada.replaceAll(Environment.getExternalStorageDirectory().getAbsolutePath(), "/sdcard");
+                        bos.write(("echo 'run_program(\"/sbin/umount\",\"/sdcard\");' >> /cache/recovery/extendedcommand\n").getBytes());
+                        bos.write(("echo 'run_program(\"/sbin/mount\",\"/sdcard\");' >> /cache/recovery/extendedcommand\n").getBytes());
+                    }
 
-                bos.write(("echo 'install_zip(\""+ fileCWM +"\");\n' >> /cache/recovery/extendedcommand\n").getBytes());
-                        /*String fileCWM2=this.romseleccionada.replaceAll(Environment.getExternalStorageDirectory().getAbsolutePath(),"/sdcard2");
-                        bos.write(("echo 'install_zip(\""+ fileCWM2 +"\");' >> /cache/recovery/extendedcommand\n").getBytes());*/
-                bos.flush();
-                bos.close();
-                rebootRecoveryQuestionFlashear();
+                    bos.write(("echo 'install_zip(\""+ fileCWM +"\");\n' >> /cache/recovery/extendedcommand\n").getBytes());
+                            /*String fileCWM2=this.romseleccionada.replaceAll(Environment.getExternalStorageDirectory().getAbsolutePath(),"/sdcard2");
+                            bos.write(("echo 'install_zip(\""+ fileCWM2 +"\");' >> /cache/recovery/extendedcommand\n").getBytes());*/
+                    bos.flush();
+                    bos.close();
+                    rebootRecoveryQuestionFlashear();
+                }else{
+                    Toast.makeText(getBaseContext(),getResources().getString(R.string.msgErrorMD5),Toast.LENGTH_SHORT).show();
+                }
             }
         } catch (Exception e) {
             Toast.makeText(getBaseContext(), getResources().getString(R.string.msgGenericError)+" 157", Toast.LENGTH_SHORT).show();
@@ -385,32 +398,36 @@ public class RomScreen extends Activity implements AdapterView.OnItemSelectedLis
     public void flashZipRecoveryOficial(){
         try {
             if(aceptadoNoModelo){
-                String application_name = "";
-                application_name = "com.mediatek.updatesystem.UpdateSystem";
-                Intent intent = new Intent("android.intent.action.MAIN");
-                List<ResolveInfo> resolveinfo_list = getPackageManager().queryIntentActivities(intent, 0);
-                boolean existe = false;
-                for (ResolveInfo info : resolveinfo_list) {
-                    if (info.activityInfo.packageName.equalsIgnoreCase("com.mediatek.updatesystem")) {
-                        if (info.activityInfo.name.equalsIgnoreCase(application_name)) {
-                            File f = new File(this.zipseleccionada);
-                            if (new File(Environment.getExternalStorageDirectory() + "/update.zip").exists()) {
-                                new File(Environment.getExternalStorageDirectory() + "/update.zip").delete();
+                if(Utilidades.checkFileMD5(new File(this.zipseleccionada))){
+                    String application_name = "";
+                    application_name = "com.mediatek.updatesystem.UpdateSystem";
+                    Intent intent = new Intent("android.intent.action.MAIN");
+                    List<ResolveInfo> resolveinfo_list = getPackageManager().queryIntentActivities(intent, 0);
+                    boolean existe = false;
+                    for (ResolveInfo info : resolveinfo_list) {
+                        if (info.activityInfo.packageName.equalsIgnoreCase("com.mediatek.updatesystem")) {
+                            if (info.activityInfo.name.equalsIgnoreCase(application_name)) {
+                                File f = new File(this.zipseleccionada);
+                                if (new File(Environment.getExternalStorageDirectory() + "/update.zip").exists()) {
+                                    new File(Environment.getExternalStorageDirectory() + "/update.zip").delete();
 
+                                }
+                                f.renameTo(new File(Environment.getExternalStorageDirectory() + "/update.zip"));
+                                Intent launch_intent = new Intent("android.intent.action.MAIN");
+                                launch_intent.setComponent(new ComponentName(info.activityInfo.packageName, info.activityInfo.name));
+                                launch_intent.setFlags(Intent.FLAG_ACTIVITY_NEW_TASK);
+                                this.startActivity(launch_intent);
+                                existe = true;
+                                break;
                             }
-                            f.renameTo(new File(Environment.getExternalStorageDirectory() + "/update.zip"));
-                            Intent launch_intent = new Intent("android.intent.action.MAIN");
-                            launch_intent.setComponent(new ComponentName(info.activityInfo.packageName, info.activityInfo.name));
-                            launch_intent.setFlags(Intent.FLAG_ACTIVITY_NEW_TASK);
-                            this.startActivity(launch_intent);
-                            existe = true;
-                            break;
                         }
                     }
+                    if (!existe) {
+                        Toast.makeText(getBaseContext(), getResources().getString(R.string.msgIngenieroNoExiste), Toast.LENGTH_SHORT).show();
+                    }
                 }
-                if (!existe) {
-                    Toast.makeText(getBaseContext(), getResources().getString(R.string.msgIngenieroNoExiste), Toast.LENGTH_SHORT).show();
-                }
+            }else{
+                Toast.makeText(getBaseContext(),getResources().getString(R.string.msgErrorMD5),Toast.LENGTH_SHORT).show();
             }
         } catch (Exception e) {
             Toast.makeText(getBaseContext(), getResources().getString(R.string.msgGenericError)+" 157", Toast.LENGTH_SHORT).show();
